Разбивает код на модули. 5дз
