+++ js/ulil.js
@@ -1,0 +1,41 @@
+//Функция, которая генерирует и возвращает случайное целое число из диапазона
+function getRandomInteger (min, max) {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+//Функция, которая генерирует и возвращает случайное уникальное целое число из диапазона
+function createRandomIdFromRangeGenerator (min, max) {
+  const previousValues = [];
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      //console.error('Перебраны все числа из диапазона от ' + min + ' до ' + max);
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+//Функция, которая возвращает случайный элемент из массива
+function randomArrayElement(array) {
+  return array[getRandomInteger(0, array.length - 1)];
+}
+
+//Функция, которая возвращает последовательно сгенерированные идентификаторы объектов, счетчик
+const createIdGenerator = () => {
+  let lastGeneratedId = 0;
+
+  return () => {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+export { getRandomInteger, createRandomIdFromRangeGenerator, randomArrayElement, createIdGenerator };
